+++ src/main.ts
@@ -1,5 +1,6 @@
 import dayjs from 'dayjs';
 import customParseFormat from 'dayjs/plugin/customParseFormat';
+import localizedFormat from 'dayjs/plugin/localizedFormat';
 
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
@@ -16,6 +17,10 @@ import Router from './router';
 import App from './App.vue';
 
 dayjs.extend(customParseFormat);
+dayjs.extend(localizedFormat);
+
+import.meta.glob('../node_modules/dayjs/locale/*.js', { eager: true });
+dayjs.locale(l10n.getLocale().toLowerCase().replace('_', '-'));
 
 const app = createApp(App);